hallbook/requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==39.0.0
Django==4.1.5
django-admin==2.0.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-db==0.0.7
django-excel-base==1.0.5
django-excel-response2==3.0.3
django-phonenumber-field==7.0.2
django-six==1.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
phonenumbers==8.13.6
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
pytz==2022.7.1
screen==1.0.1
sqlparse==0.4.3
xlwt==1.3.0asgiref==3.6.0
cffi==1.15.1
cryptography==39.0.0
Django==4.1.5
django-admin==2.0.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-db==0.0.7
django-excel-base==1.0.5
django-excel-response2==3.0.3
django-phonenumber-field==7.0.2
django-six==1.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
phonenumbers==8.13.6
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
pytz==2022.7.1
screen==1.0.1
sqlparse==0.4.3
xlwt==1.3.0
